Remove indexes before removing a column.
The migration tests indicate that we should be able to remove columns that have indexes on them. SQLServer doesn't like that, so we'll attempt to remove an indexes on this column before we remove the column.  Had to reorg the indexes method into a private a public one.  The public one messes with the autocommit status of the connection for somereason and that was breaking the remove_column call.  Maybe the public one doesn't need to do this though.