Change how we handle limits for columns and how we store @limit for making schema_dumper_test work.
We now store @limit as 2 for smallint and 8 for bigint (so schema dumper will output :integer, :limit => 2 or 8 and we'll get smallint or bigint columns created accordingly.  We also say that if limit is <3 it's a smallint and >4 it's a bigint, before you had to be limit ==4 to be an integer and limit ==3 would be a smallint, which is not quite correct. (Limit of 3 is bigger than a smallint).