Get rid of a bunch of tests that fail because of differences in the number of quries AR expects it to run.
We have to open up ActiveRecord::Base.connection in our (new) sqlserver_helper and add a regexp for SELECT SCOPE_IDENTITY to it.  This means that assert_queries will ignore any queries that match this regexp and so the tests pass.  Or, to put it more correctly, execute_with_query_record won't record them.