Provided alternative length of specific large data types since COL_LENGTH returns -1 and 16 for some large data types.
Previously, this would cause issues when ActiveRecord used this length to
check if data can fit into this column, such as sessions.data.

The specific error I encountered:
'Your session data is larger than the data column in which it is to be stored. You must increase the size of your data column if you intend to store large data.'