New simple 2012 schema.
The idea is that we want a single [datatypes] table to put every type that SQL Sever 2012 understands. This will have a signle test model and we use this models columns to test our new ActiveRecord::Type objects. This current 2012 list might be incomplete. If so, we will add a test. Any newer datatypes for 2014 will be added to a unique 2014.sql file.

* Rename test/models_sqlserver to test/models/sqlserver. No files moved.
* Deprecate every single model/#{name}.rb. We will only be moving what we need.
* New models in test/models/sqlserver MUST use a `SS` prefix for the model name!!!