Remove all auto recoonect & SQL retry logic.
This was getting too complicated and I believe it stood in the way of passing a few tests along with the strong likelyhood of causing other errors. For example, I found out that using the super strong method of going into single DB user mode caused potential DB lock outs and other grief when used with our connection pool #reset! hook. I would not be surprised to find out that we played a hand in causing users grief in odd to find places when we tried doing to much for them. Lets just allow ActiveRecord and the connection pool to do what it needs and if we are no longer #active? then so be it.

🏃🔥 So passes Autoreconnectathor, son of Notourproblemalion.