[Rails5] Ensure date/time columns are not nil on create/update.
Creating or updating records resets the attributes by mapping the hash
using ActiveRecord's `Attribute#forgetting_assignment` method. This
method essentiall sets mimic the reload by setting the attributes to
FromDatabase using the FromUser value "for the database".

This is a problem for us because the `value_for_database` for datetime
objects are strings that can not be re-parsed due to SQL Server's time
formats. For example, a date of August 18th 2016 would be formatted as
'08-19-2016' and fail deserialize:

  ActiveModel::Type::Date.new.deserialize('08-19-2016') # => nil

Both ActiveModel's date and time types rescue `Date.parse` methods of
`new_date` and `new_time` with nil. It was suggested that we use
Data classes for our time types and we may explore that after this
commit which essentially prepends a new `forgetting_assignment` method
that checks the type. If it is a SQL Server date/time, it will just
convert that value object to a FromDatabase using the existing value.