[Rails 5] Fix some times coming back nil after save.
As part of #528, a few errors were introduced. Since we have a `fast_string_to_time` implementation that uses a custom format which includes %N, parsing the time for DB from DB could fail. So we ensure we always have a nanosecond on the SQL even if it is 0. No need not to.