Fix deteministic fetch when table has a composite primary key
Even though Rails does not support tables with primary keys, it still allows
those tables to be defined without any problems.

The SQL Server adapter will always try to retrieve records in a deterministic
way but it currently fails if the table has a composite primary key, which
shouldn't happen.

This PR fixes how the adapter determines which column (generally the PK) is
used as the default order when multiple primary keys are detected.

It first checks if one of the primary keys is also an identity column. If
there's an identity column, it will choose it as the "main primary key",
otherwise it will just fallback to the first of the primary keys.