Reimplement test that build queries with group by and then call
Rails introduced the ability to call delete_all and update_all on scopes
that use having and group by doing a subquery. See
[here](https://github.com/rails/rails/commit/989d53ba3b3b43d3b1a179a146be9244e1c2537e)
and
[here](https://github.com/rails/rails/commit/4acb6660e265fd40b38e301fbd15909574d6a59e)

To test this they build scopes like
`Post.group('posts.id').joins(:comments).having('count('comments.id') >
2').all`. SQL Server adapter doesn't support these. Results in
```
ActiveRecord::StatementInvalid: TinyTds::Error: Column 'posts.author_id'
is invalid in the select list because it is not contained in either an
aggregate function or the GROUP BY clause.
```

Coerce and reimplement the test to use subqueries.