Reverse order of values when upserting
I contributed two tests to Rails because I was not sure how `insert_all` and `upsert_all` should behave with duplicate identifiers.

For `insert_all`, all DBMS only insert the first value. For `upsert_all`, PostgreSQL throws an error in this scenario, but with sqlite and MySQL, the last entry ends up in the database.

For now, I do not have a smarter idea than reversing the `inserts`, which ultimately provide the `values_list`. Adding `MAX` would be another, but I fear the SQL query would get much more complicated.