Fix upserting with mixed data types
The previous `CASE` statement used `COALESCE` with `'NULL'` as a fallback value if either `source` or `target` were `NULL`. However, `'NULL'` as written in this statement, is a `varchar` and therefore does not work when comparing with other datatypes.

This commit changes the switch statement to check if either both values are equal, or both values are actually `NULL`.